Moved demo poject, Renaming demo project, moved assembly, dissabled xml warnings in demo, splited EventBusDocUIOptioins
