+++ BusEventDoc.sln
@@ -17,9 +17,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kros.EventBusDoc.UI", "src\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kros.EventBusDoc.Demo", "demo\Kros.EventBusDoc.Demo\Kros.EventBusDoc.Demo.csproj", "{3A630BA1-4F38-4988-B74B-F6099EEB08F1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kros.EventBusDoc.Demo.Services", "Kros.EventBusDoc.Demo.MessageService\Kros.EventBusDoc.Demo.Services.csproj", "{ABE3486E-EB78-456C-A438-1E37EDCD7CBD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kros.EventBusDoc.Demo2", "demo\Kros.EventBusDoc.Demo2\Kros.EventBusDoc.Demo2.csproj", "{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kros.EventBusDoc.Demo2", "demo\Kros.EventBusDoc.Demo2\Kros.EventBusDoc.Demo2.csproj", "{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kros.EventBusDoc.Demo.Contracts", "demo\Kros.EventBusDoc.Demo.MessageService\Kros.EventBusDoc.Demo.Contracts.csproj", "{75EABEEE-CAF0-4E30-B114-300226F1E83E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,14 +43,14 @@ Global
 		{3A630BA1-4F38-4988-B74B-F6099EEB08F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3A630BA1-4F38-4988-B74B-F6099EEB08F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3A630BA1-4F38-4988-B74B-F6099EEB08F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABE3486E-EB78-456C-A438-1E37EDCD7CBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABE3486E-EB78-456C-A438-1E37EDCD7CBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABE3486E-EB78-456C-A438-1E37EDCD7CBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABE3486E-EB78-456C-A438-1E37EDCD7CBD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{28BC2F6D-9F4D-4268-ADC8-161C31B7366A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{75EABEEE-CAF0-4E30-B114-300226F1E83E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{75EABEEE-CAF0-4E30-B114-300226F1E83E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{75EABEEE-CAF0-4E30-B114-300226F1E83E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{75EABEEE-CAF0-4E30-B114-300226F1E83E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -60,8 +60,8 @@ Global
 		{057EC6FF-601A-4997-836E-476A096A200A} = {F3D1886D-3C97-4435-B872-B036BD6F604C}
 		{EC17EF90-83F6-476C-9D97-1CE2E3069ADC} = {F3D1886D-3C97-4435-B872-B036BD6F604C}
 		{3A630BA1-4F38-4988-B74B-F6099EEB08F1} = {00775576-34FF-429F-A72A-EC9C956A3F7D}
-		{ABE3486E-EB78-456C-A438-1E37EDCD7CBD} = {00775576-34FF-429F-A72A-EC9C956A3F7D}
 		{28BC2F6D-9F4D-4268-ADC8-161C31B7366A} = {00775576-34FF-429F-A72A-EC9C956A3F7D}
+		{75EABEEE-CAF0-4E30-B114-300226F1E83E} = {00775576-34FF-429F-A72A-EC9C956A3F7D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {64F1C94B-4EEA-4398-8EE3-274E4E713BB9}
