+++ demo/Kros.EventBusDoc.Demo.MessageService/Kros.EventBusDoc.Demo.Contracts.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Kros.EventBusDoc.Demo.Contracts</AssemblyName>
+    <RootNamespace>Kros.EventBusDoc.Demo.Contracts</RootNamespace>
+    <NoWarn>1701;1702;1591</NoWarn>
+  </PropertyGroup>
+
+</Project>
