+++ demo/Kros.EventBusDoc.Demo/Kros.EventBusDoc.Demo.csproj
@@ -7,7 +7,7 @@
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
     <DocumentationFile>C:\Develop\bus\demo\Kros.EventBusDoc.Demo\Kros.EventBusDoc.Demo.xml</DocumentationFile>
-    <NoWarn>1701;1702;1591</NoWarn>
+    <NoWarn>1701;1702;1591;1998</NoWarn>
   </PropertyGroup>
 
 
@@ -19,9 +19,9 @@
 
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Kros.EventBusDoc.Demo.MessageService\Kros.EventBusDoc.Demo.Services.csproj" />
     <ProjectReference Include="..\..\src\Kros.EventBusDoc.Generator\Kros.EventBusDoc.Generator.csproj" />
     <ProjectReference Include="..\..\src\Kros.EventBusDoc.UI\Kros.EventBusDoc.UI.csproj" />
+    <ProjectReference Include="..\Kros.EventBusDoc.Demo.MessageService\Kros.EventBusDoc.Demo.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
