+++ demo/Kros.EventBusDoc.Demo2/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-﻿using Kros.EventBusDoc.Demo.Services.Events;
+﻿using Kros.EventBusDoc.Demo.Contracts.Events;
 using Kros.EventBusDoc.Generator.BusentAnnotation;
 
 [assembly: EventBusEvent(EventType = typeof(IMessageEvent))]
