+++ src/Kros.EventBusDoc.UI/Middleware/Options/UrlDescriptor.cs
@@ -1,0 +1,18 @@
+﻿namespace Kros.EventBusDoc.UI.Middleware.Options
+{
+    /// <summary>
+    /// Description for endpoint with url and name
+    /// </summary>
+    public class UrlDescriptor
+    {
+        /// <summary>
+        /// Endpoint's Url
+        /// </summary>
+        public string Url { get; set; }
+
+        /// <summary>
+        /// Endpoint's name
+        /// </summary>
+        public string Name { get; set; }
+    }
+}
